# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FinançasPro API",
    "version": "11.0.0",
    "description": "API SaaS de finanças pessoais. Corpos de request gerados a partir dos schemas Zod que a API valida em runtime."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Listar transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      },
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Criar transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "receita",
                      "despesa",
                      "transferencia"
                    ]
                  },
                  "amount": {
                    "type": "number",
                    "maximum": 999999999999.99,
                    "exclusiveMinimum": 0
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "category": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "subcategory": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "accountId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "targetAccountId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 40
                    },
                    "maxItems": 10,
                    "default": []
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true
                  },
                  "recurring": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "type",
                  "amount",
                  "description",
                  "category",
                  "date"
                ]
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Obter transactions por id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Atualizar transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "receita",
                      "despesa",
                      "transferencia"
                    ]
                  },
                  "amount": {
                    "type": "number",
                    "maximum": 999999999999.99,
                    "exclusiveMinimum": 0
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "category": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "subcategory": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "accountId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "targetAccountId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 40
                    },
                    "maxItems": 10,
                    "default": []
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true
                  },
                  "recurring": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Excluir transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/billing/{orgId}/subscription": {
      "get": {
        "summary": "Assinatura atual da organização",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/{orgId}/subscribe": {
      "post": {
        "summary": "Assinar um plano",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/{orgId}/checkout": {
      "post": {
        "summary": "Abrir sessão de checkout do Stripe",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/{orgId}/portal": {
      "post": {
        "summary": "Abrir o portal de cobrança do Stripe",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/{orgId}/cancel": {
      "post": {
        "summary": "Cancelar ao fim do período vigente",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/{orgId}/resume": {
      "post": {
        "summary": "Reativar renovação (desfaz o cancelamento agendado)",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/billing/{orgId}/invoices": {
      "get": {
        "summary": "Faturas emitidas",
        "tags": [
          "Billing"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/open-finance/providers": {
      "get": {
        "summary": "Provedores disponíveis (Belvo, sandbox)",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/open-finance/connections": {
      "get": {
        "summary": "Conexões bancárias do usuário",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Criar conexão bancária",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/open-finance/connections/{id}": {
      "delete": {
        "summary": "Remover conexão bancária",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/open-finance/connections/{id}/sync": {
      "post": {
        "summary": "Importar lançamentos — deduplica por openFinanceId",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/open-finance/belvo/widget-token": {
      "post": {
        "summary": "Token de acesso ao widget da Belvo",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/open-finance/belvo/complete": {
      "post": {
        "summary": "Concluir o vínculo após o widget",
        "tags": [
          "Open Finance"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}